Refactored the build, putting a bunch of logic in grunt/sim-manager.js.  I also made way for the logic that decides what to build based on build timestamps, but I still need to create that functionality.
