+++ package.json
@@ -28,6 +28,7 @@
     "grunt-gh-pages": "^0.9.1",
     "grunt-targethtml": "^0.2.6",
     "matchdep": "^0.3.0",
+    "underscore": "^1.8.3",
     "wrench": "^1.5.8"
   }
 }
